// apvue-admin/src/api/admin/models/banner.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 轮播图
 */
export type banner = {
    id?: number;
    title?: string;
    imgurl?: string;
    type?: string;
    targetVideoId?: number;
    targetTopicId?: number;
    turl?: string;
    sort?: number;
    status?: number;
    createdat?: string;
};