// apvue-admin/src/api/admin/models/series.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 剧集
 */
export type series = {
    id?: number;
    title?: string;
    description?: string;
    coverUrl?: string;
    cid?: string;
    status?: string;
    ecount?: string;
    createdtime?: string;
    updatedtime?: string;
    file?: Blob;
};